+++ x/oracle/keeper/hooks.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	perptypes "github.com/NibiruChain/nibiru/x/perp/types/v1"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
@@ -8,7 +9,6 @@ import (
 	oracletypes "github.com/NibiruChain/nibiru/x/oracle/types"
 
 	"github.com/NibiruChain/nibiru/x/epochs/types"
-	perptypes "github.com/NibiruChain/nibiru/x/perp/types"
 )
 
 var _ types.EpochHooks = Hooks{}